+++ client/src/components/Navigation/PrivateRoute.js
@@ -2,6 +2,7 @@ import React from "react";
 import { Router, Switch, Route } from "react-router-dom";
 import Home from '../Home';
 import history from './history';
+import Landing from '../Landing'
 
 export default function PrivateRoute({
   //authenticated,
@@ -11,7 +12,7 @@ export default function PrivateRoute({
 
     <Router history={history}>
       <Switch>
-      <Route path="/" exact component={Home} />
+      <Route path="/" exact component={Landing} />
       </Switch>
     </Router>
   );
