+++ config.js
@@ -1,8 +1,8 @@
 let config = {
-  host    : 'ov-research-4.uwaterloo.ca',
-  user    : '',
-  password: '',
-  database: ''
+  host    : 'ec2-18-216-101-119.us-east-2.compute.amazonaws.com',
+  user    : 'krajesh',
+  password: 'J@yd33p1820',
+  database: 'krajesh'
 };
  
 module.exports = config;
